src/python/data_recorder/pyproject.toml
[project]
name = "data-recorder"
version = "0.1.0"
description = "Data recorder node for system identification"
requires-python = ">=3.11"
dependencies = [
    "dora-rs==0.4.1",
    "pyarrow>=23.0.1",
    "numpy",
]
